VK: Remove explicitImageReadyWait callback (#8603)
This function was created to make sure the swapchain
image was available before submitting the work to the GPU.

So the order of the process looked like

| Acquire | Begin | Record | Wait | Present | End |

After this change the expectation is that when the
swapchain acquires a new image it will also wait until
is actually available from the platform.

| Acquire | Begin | Record | Present | End |

In the case the image is not available, the expectation
is that the acquire will wait until it is and then start
processing the frame.