VK: Fix leak in VulkanExternalImageManager (#9194)
In the cases of a descriptor set being updated
with a external sampled texture but that descriptor
set doesn't get used later on for rendering.

Their corresponding entry in VulkanExternalImageManager::mSetBindings
will never get removed until the application is terminated, causing
performance issues or a fatal crash if run long enough.

The expectation is that the amount of bindings with
external samplers is low or none, so this won't cause
any regression when not using external samplers.